% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes 0.0.1.R
\name{custom_theme_options}
\alias{custom_theme_options}
\title{Change colors and theme options for pprint}
\usage{
custom_theme_options(
  primary_col = "#00CC66",
  secondary_col = "red",
  tertiary_col = "blue",
  check_for_Rstudio_theme = TRUE
)
}
\arguments{
\item{primary_col}{change default primary color (the color used when not using colors from Rstudio themes)}

\item{secondary_col}{change default secondary color (the color used when not using colors from Rstudio themes)}

\item{tertiary_col}{change default tertiary color (the color used when not using colors from Rstudio themes)}

\item{check_for_Rstudio_theme}{if = FALSE pprint never uses colors gathered from Rstudio theme}
}
\description{
Change colors and theme options for pprint
}
