% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{pfactor}
\alias{pfactor}
\title{Summary for fa/principal object from psych package}
\usage{
pfactor(
  factor_model,
  primary = theme_colors(primary = T),
  secondary = theme_colors(secondary = T),
  plaintext = theme_colors(fg = T),
  threshold = 0.4,
  eigenvalues = T,
  highlight_correlations = 0.2,
  pattern_matrix = T,
  correlations = T,
  rotation_map = F,
  scree_plot = F,
  descriptives = F
)
}
\arguments{
\item{factor_model}{pscyh fa/principal object}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}

\item{threshold}{factorloadings below threshold are blanked default = 0.4}

\item{eigenvalues}{if TRUE (default) eigenvalues are printed}

\item{highlight_correlations}{factor correlations above the value are highlighted}

\item{correlations}{if TRUE (default) factor correlations are printed}

\item{rotation_map}{if TRUE (default) rotation map is printed}

\item{scree_plot}{if TRUE screeplot is printed NB: not implemented yet}

\item{descriptives}{if TRUE descriptives are included NB: not implemented yet}

\item{Pattern_Matrix}{if TRUE (default) pattern matrix is printed}
}
\description{
Summary for fa/principal object from psych package
}
