% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprint_logistic_regression.R
\name{plogit}
\alias{plogit}
\title{Summary for logistic regression model using GLM}
\usage{
plogit(logistic_reg_model, odds_ratio = FALSE, accuracy = FALSE, ...)
}
\arguments{
\item{logistic_reg_model}{a GLM type object family = binomial(link = "logit)}

\item{accuracy}{If = TRUE TRUE/FALSE positives/negatives are included, and specificity, sensitivity and positiv/negative predictive values}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}
}
\description{
Summary for logistic regression model using GLM
}
