% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style and summary.R
\name{pprint}
\alias{pprint}
\title{pretty print for dataframes}
\usage{
pprint(
  x,
  primary = theme_colors(primary = T),
  secondary = theme_colors(secondary = T),
  plaintext = theme_colors(fg = T),
  style = "default",
  rows = 20
)
}
\arguments{
\item{x}{a dataframe to be printed in pprint style}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}

\item{style}{chose style to print in, currently only "default" is supported}

\item{rows}{maximum number of rows to print}
}
\description{
pretty print for dataframes
}
