% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preg.R
\name{preg}
\alias{preg}
\title{Summary for linear regressions}
\usage{
preg(
  x,
  primary = theme_colors(primary = T),
  secondary = theme_colors(secondary = T),
  plaintext = theme_colors(fg = T),
  std_beta = FALSE,
  ...
)
}
\arguments{
\item{x}{lm-object}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}

\item{std_beta}{if TRUE std.beta for continuous variables are included, or cohens'd for factor variables}

\item{...}{}
}
\description{
Summary for linear regressions
}
