% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style and summary.R
\name{psummary.data.frame}
\alias{psummary.data.frame}
\title{Summary for dataframes USE psummary() instead}
\usage{
\method{psummary}{data.frame}(x, haven_labelled = FALSE, rows = nrow(x), ...)
}
\arguments{
\item{x}{dataframe, matrix or tibble object}

\item{rows}{maximum number of rows to print}

\item{...}{arguments passed to functions within the function}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}

\item{style}{chose style to print in, currently only "default" is supported}
}
\description{
Summary for dataframes USE psummary() instead
}
