% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprint_logistic_regression.R
\name{psummary.glm}
\alias{psummary.glm}
\title{Summary for logistic regressions}
\usage{
\method{psummary}{glm}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Summary for logistic regressions
}
