% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psummary_lm.R
\name{psummary.lm}
\alias{psummary.lm}
\title{Summary for linear regressions}
\usage{
\method{psummary}{lm}(x, std_beta = FALSE, standardizeAll = FALSE, ...)
}
\arguments{
\item{x}{lm-object}

\item{std_beta}{if TRUE std.beta for continuous variables are included, or cohens'd for factor variables}

\item{...}{}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}
}
\description{
Summary for linear regressions
}
