% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style and summary.R
\name{psummary}
\alias{psummary}
\title{Summary for different types of objects}
\usage{
psummary(object, ...)
}
\arguments{
\item{object}{a dataframe/tibble/matrix, lm-object, psych fa/principal object, or GLM-object family = binomial(link = "logit")}

\item{...}{arguments passed on to the summary functions for each object, respectively: pdataframe, preg, pfactor and plogit}
}
\description{
Summary for different types of objects
}
