% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style and summary.R
\name{pdataframe}
\alias{pdataframe}
\title{Summary for dataframes USE psummary instead}
\usage{
pdataframe(
  x,
  primary = theme_colors(primary = T),
  secondary = theme_colors(secondary = T),
  plaintext = theme_colors(fg = T),
  style = "default",
  rows = nrow(x),
  ...
)
}
\arguments{
\item{x}{dataframe, matrix or tibble object}

\item{primary}{primary colour this is automatically chosen, but can be overwritten}

\item{secondary}{secondary colour this is automatically chosen, but can be overwritten}

\item{plaintext}{colour for plaintext this is automatically chosen, but can be overwritten}

\item{style}{chose style to print in, currently only "default" is supported}

\item{rows}{maximum number of rows to print}

\item{...}{arguments passed to functions within the function}
}
\description{
Summary for dataframes USE psummary instead
}
